Move authenticate.hook.ts and login-required.hook.ts
