+++ packages/core/src/auth/authentication/index.ts
@@ -1,4 +1,3 @@
 export * from './strategies';
-export * from './authenticate.hook';
 export * from './authenticator.interface';
 export * from './login.controller';
