+++ packages/core/src/auth/authorization/index.ts
@@ -1,2 +1,1 @@
 export { PermissionRequired } from './permission-required.hook';
-export { LoginRequired } from './login-required.hook';
