+++ packages/core/src/auth/index.ts
@@ -2,3 +2,5 @@ export * from './authentication';
 export * from './authorization';
 export * from './entities';
 export * from './utils';
+export * from './authenticate.hook';
+export { LoginRequired } from './login-required.hook';
